replace singular eventual-factory with a per call factory
