+++ packages/@eventual/core-runtime/src/workflow/workflow-executor.ts
@@ -2,7 +2,6 @@ import {
   DeterminismError,
   SystemError,
   WorkflowTimeout,
-  type Signal,
   type Workflow,
   type WorkflowContext,
 } from "@eventual/core";
@@ -17,16 +16,15 @@ import {
   isWorkflowStarted,
   isWorkflowTimedOut,
   iterator,
+  type Call,
   type CallEvent,
   type CompletionEvent,
-  type Call,
   type EventualHook,
   type EventualPromise,
-  type Property,
-  type PropertyType,
   type HistoryEvent,
   type HistoryStateEvent,
-  type SignalReceived,
+  type Property,
+  type PropertyType,
   type WorkflowCallHistoryEvent,
   type WorkflowInputEvent,
   type WorkflowRunStarted,
@@ -34,15 +32,30 @@ import {
   type _Iterator,
 } from "@eventual/core/internal";
 import { isPromise } from "util/types";
-import { createEventualFromCall } from "../eventual-factory.js";
 import { enterEventualCallHookScope } from "../eventual-hook.js";
-import { formatExecutionId } from "./execution.js";
-import { isFailed, isResolved, isResult } from "../result.js";
-import { filterEvents } from "./events.js";
 import {
   PropertyRetriever,
   getEventualProperty,
 } from "../property-retriever.js";
+import { isFailed, isResolved, isResult } from "../result.js";
+import { filterEvents } from "./events.js";
+import {
+  isAfterEveryEventTrigger,
+  isEventTrigger,
+  isPromiseTrigger,
+  isResolvedEventualDefinition,
+  isSignalTrigger,
+  type AfterEveryEventTrigger,
+  type EventTrigger,
+  type SignalTrigger,
+  type TriggerHandler,
+  type UnresolvedEventualDefinition,
+} from "./eventual-definition.js";
+import { formatExecutionId } from "./execution.js";
+import {
+  EventualFactory,
+  createDefaultEventualFactory,
+} from "./call-eventual-factory.js";
 
 /**
  * Put the resolve method on the promise, but don't expose it.
@@ -203,7 +216,8 @@ export class WorkflowExecutor<Input, Output, Context = undefined> {
     private workflow: Workflow<any, Input, Output>,
     public history: HistoryStateEvent[],
     // TODO: properties used in the workflow should be encoded in the history to keep them constant between runs
-    private propertyRetriever: PropertyRetriever
+    private propertyRetriever: PropertyRetriever,
+    private eventualFactory: EventualFactory = createDefaultEventualFactory()
   ) {
     this.nextSeq = 0;
     this.expected = iterator(history, isCallEvent);
@@ -452,7 +466,7 @@ export class WorkflowExecutor<Input, Output, Context = undefined> {
         }
 
         try {
-          const eventual = createEventualFromCall(call);
+          const eventual = self.eventualFactory.createEventualDefinition(call);
           const seq = self.nextSeq++;
 
           // not all calls generate events, these calls will not emit calls and thus will not cause determinism errors.
@@ -785,140 +799,6 @@ export interface WorkflowResult<T = any> {
    * This can still be non-empty even if the chain has terminated because of dangling promises.
    */
   calls: WorkflowCall[];
-}
-
-export type Trigger<Output> =
-  | PromiseTrigger<Output>
-  | EventTrigger<Output>
-  | AfterEveryEventTrigger<Output>
-  | SignalTrigger<Output>;
-
-export const Trigger = {
-  onPromiseResolution: <Output = any, Input = any>(
-    promise: Promise<Input>,
-    handler: PromiseTrigger<Output, Input>["handler"]
-  ): PromiseTrigger<Output, Input> => {
-    return {
-      promise,
-      handler,
-    };
-  },
-  afterEveryEvent: <Output = any>(
-    handler: AfterEveryEventTrigger<Output>["afterEvery"]
-  ): AfterEveryEventTrigger<Output> => {
-    return {
-      afterEvery: handler,
-    };
-  },
-  onWorkflowEvent: <Output, T extends CompletionEvent["type"]>(
-    eventType: T,
-    handler: TriggerHandler<[event: CompletionEvent & { type: T }], Output>
-  ): EventTrigger<Output, CompletionEvent & { type: T }> => {
-    return {
-      eventType,
-      handler,
-    };
-  },
-  onSignal: <Output = any, Payload = any>(
-    signalId: Signal<Payload>["id"],
-    handler: SignalTrigger<Output, Payload>["handler"]
-  ): SignalTrigger<Output, Payload> => {
-    return {
-      signalId,
-      handler,
-    };
-  },
-};
-
-type TriggerHandler<Args extends any[], Output> =
-  | {
-      (...args: Args): void | undefined | Result<Output>;
-    }
-  | Result<Output>;
-
-export interface PromiseTrigger<Output, Input = any> {
-  promise: Promise<Input>;
-  handler: TriggerHandler<[val: Result<Input>], Output>;
-}
-
-export interface AfterEveryEventTrigger<Output> {
-  afterEvery: TriggerHandler<[], Output>;
-}
-
-export interface EventTrigger<
-  out Output = any,
-  E extends CompletionEvent = any
-> {
-  eventType: E["type"];
-  handler: TriggerHandler<[event: E], Output>;
-}
-
-export interface SignalTrigger<Output, Payload = any> {
-  signalId: Signal["id"];
-  handler: TriggerHandler<[event: SignalReceived<Payload>], Output>;
-}
-
-export function isPromiseTrigger<Output, R>(
-  t: Trigger<Output>
-): t is PromiseTrigger<Output, R> {
-  return "promise" in t;
-}
-
-export function isAfterEveryEventTrigger<Output>(
-  t: Trigger<Output>
-): t is AfterEveryEventTrigger<Output> {
-  return "afterEvery" in t;
-}
-
-export function isEventTrigger<Output, E extends CompletionEvent>(
-  t: Trigger<Output>
-): t is EventTrigger<Output, E> {
-  return "eventType" in t;
-}
-
-export function isSignalTrigger<Output, Payload = any>(
-  t: Trigger<Output>
-): t is SignalTrigger<Output, Payload> {
-  return "signalId" in t;
-}
-
-interface EventualDefinitionBase {
-  /**
-   * Return the event form of a call.
-   *
-   * If undefined, the call will not be emitted by the workflow.
-   * If there was a expected event and these events do not match, the workflow will throw a {@link DeterminismError}
-   */
-  createCallEvent?: (seq: number) => WorkflowCallHistoryEvent;
-}
-
-export interface ResolvedEventualDefinition<R> extends EventualDefinitionBase {
-  /**
-   * When provided, immediately resolves an EventualPromise with a value or error back to the workflow.
-   *
-   * Commands can still be emitted, but the eventual cannot be triggered.
-   */
-  result: Result<R>;
-}
-
-export interface UnresolvedEventualDefinition<R>
-  extends EventualDefinitionBase {
-  /**
-   * Triggers give the Eventual an opportunity to resolve themselves.
-   *
-   * Triggers are only called when an eventual is considered to be active.
-   */
-  triggers: Trigger<R> | (Trigger<R> | undefined)[];
-}
-
-export type EventualDefinition<R> =
-  | ResolvedEventualDefinition<R>
-  | UnresolvedEventualDefinition<R>;
-
-export function isResolvedEventualDefinition<R>(
-  eventualDefinition: EventualDefinition<R>
-): eventualDefinition is ResolvedEventualDefinition<R> {
-  return "result" in eventualDefinition;
 }
 
 interface TriggerHandlerRef<Args extends any[]> {
